fix elf ws now returns uint64 (#2690)
* fix elf ws now returns uint64

since 0.46, elf watershed returns uint64. As we do some further processing with
vigra we need to cast to uint32 (uint64 not supported by vigra)

+++ ilastik/applets/wsdt/opWsdt.py
@@ -10,7 +10,7 @@ import vigra
 from lazyflow.utility import OrderedSignal
 from lazyflow.request import Request
 from lazyflow.graph import Operator, InputSlot, OutputSlot
-from lazyflow.roi import roiToSlice, sliceToRoi
+from lazyflow.roi import roiToSlice
 from lazyflow.operators import OpBlockedArrayCache, OpMetadataInjector
 from lazyflow.operators.generic import OpPixelOperator
 from lazyflow.utility.timer import Timer
@@ -105,6 +105,9 @@ def parallel_watershed(
             f"processing block {block_index} {block.outerBlock.begin}-{block.outerBlock.end} took {btimer.seconds()}"
         )
 
+        # elf started returning uint64 in 0.46. Casting here is not dangerous.
+        # Up to now vigra is still used to produce the watershed in elf internally.
+        ws_outer = ws_outer.astype("uint32")
         ws_inner = vigra.analysis.labelMultiArray(ws_outer[inner_local_slicing])
 
         labels[inner_slicing] = ws_inner
@@ -223,7 +226,9 @@ class OpWsdt(Operator):
                 self.ApplyNonmaxSuppression.value,
             )
 
-        result[..., 0] = ws
+        # elf started returning uint64 in 0.46. Casting here is not dangerous.
+        # Up to now vigra is still used to produce the watershed in elf internally.
+        result[..., 0] = ws.astype("uint32")
 
         self.watershed_completed()
